### 2025-07-31 #8686 @kyleniemeyer
@bahremsd a few initial notes:
- I do not see any clear installation instructions in the README; please add, even if as simple as `pip install`
- The README is quite long. Have you considered building a dedicated documentation website for this?
- The paper is currently too long, please trim, aiming for 1000 words as a target. Novel research results don't belong in the paper, and other things may belong in documentation (such as detailed description of theory/methods).
