### 2025-08-14 #8686 @bahremsd
Dear Dr. Niemeyer,

Thank you for your feedback on our JOSS submission. Below is a summary of the changes made in response to your initial comments.

---

## **Installation**
We have added an **Installation** section to the manuscript (README.md) with the following content:

> TMMax can be readily installed from the Python Package Index using
> ```bash
> pip install tmmax
> ```
> This command automatically handles all dependencies.

---

## **README Length Reduction**
- Reduced the **README** word count from **3819** to **980** words.
- Main changes:
  - Significantly condensed the **Statement of Need** and **Benchmark** descriptions.
  - **Removed** the “Code Structure” section and figure entirely.
- The removed detailed descriptions will be added to a dedicated documentation site (see next section).

---

## **Documentation Website**
- A **Read the Docs** site will be created for TMMax.
- The removed sections (e.g., extended theory, detailed method explanations, code structure) will be migrated there.
- This will ensure the README remains concise while providing full technical depth in the documentation.

---

## **Recommended Reviewers**
We propose the following reviewers for our manuscript:

- Martín Caldarola (@caldarolamartin)
- Thomas Else (@tomelse)
- Jack Tyson (@JackTyson)
- Rishabh Garg (@CrazeXD)
- Kareem H. El-Safty (@kareem1925)
- John Steven Dodge (@jsdodge)
- Jaren Nicholas Ashcraft (@Jashcraf)
- Chung-You Shih (@ldes89150)

---

Please let us know if further adjustments are needed before moving to the next stage. We appreciate your time and guidance in improving our submission.

Best regards,

Bahrem S. Danis
