### 2025-08-15 #8686 @bahremsd
Dear Dr. Niemeyer @kyleniemeyer ,

We discovered a small typo in the text and corrected it. The most recent PDF version, generated via editorbot, is provided above.

Best regards,

Bahrem S. Danis
