### 2025-08-19 #8686 @kyleniemeyer
@bahremsd excellent, thanks. I see the link now in the GitHub repo; I'd just suggest adding a link to the "Documentation" section in the README (right now that directs to the `docs` directory).

Unfortunately, we do not have an editor free to handle your submission right now, so I have to place this on our waitlist until someone becomes available.
