### 2025-09-10 #8686 @bahremsd
Hi @kyleniemeyer ,

Thank you for the suggestion. We have updated the README and added the following sentence under the "Documentation" section:

“The complete documentation for TMMax is available in the [TMMax Read The Docs](https://tmmax.readthedocs.io/en/latest/) .”

Best regards,
Bahrem S. Danis
