### 2025-09-18 #8686 @lockwo
> [@lockwo](https://github.com/lockwo) Sure! As this is my first time reviewing, is there any guidance I need to look over?

Once we have other reviewers, I'll make the official review issue (which will be a new thread), but before then you can checkout the reviewing docs on the JOSS website: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. There's a few pages that outline the ideas of JOSS reviewing (and how it is different than other "typical" academic journals)
