### 2025-09-27 #9088 @editorialbot
[REVIEW]: TMMax: High-performance modeling of multilayer thin-film structures using transfer matrix method with JAX
**Submitting author:** @bahremsd (<a href="http://orcid.org/0009-0002-9880-0446">BAHREM SERHAT DANIS</a>)
**Repository:** https://github.com/bahremsd/tmmax
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.4
