### 2025-09-30 #9088 @bahremsd
Response to Reviewers

Dear Editors and Reviewers,

We would like to thank you for accepting the revision of our paper. We are especially grateful to Reviewer Rishabh Garg for the thoughtful and valuable comments. Below, we provide detailed responses to each point, along with the corresponding actions taken.

---

1. Reviewer comment:
“My first suggestion is for your paper to make the target audience of this software should be made more clear. Who would be using TMMax? Optical engineers? In what specific fields? Just 1 sentence should be enough for this.”

1. Our response:
We appreciate this valuable comment. TMMax is designed to support all researchers working on thin-film design, particularly those dealing with complex multilayer stacks. The software facilitates simulations of thin films with complex structures. Therefore, TMMax can be useful for optical engineers as well as for researchers in optics and photonics who are engaged in dielectric thin-film design.

1. Action taken:
We revised the Summary section to make the target audience more explicit. Specifically, we removed the sentence:
> “providing a significant advantage in computational efficiency. Our method enables rapid and scalable simulations of large and complex multilayer thin-film structures, significantly accelerating the design process.”

and replaced it with the following, which clarifies the intended users without exceeding the 1000-word limit:
> “Our method enables optical engineers and thin-film researchers in optics and photonics to efficiently design complex dielectric multilayer structures through rapid and scalable simulations.”

---

2. Reviewer comment:
“Referencing openjournals/joss-reviews#9088 (comment), I would recommend adding the DOI for those that editorial bot found to be missing despite likely existing. While I did confirm that the DOI's that it found existed, if any are incorrect, feel free to mention.”

2. Our response:
We carefully reviewed all references flagged by the editorial bot. DOIs were added wherever available. However, for some references that are GitHub repositories rather than published articles, DOIs do not exist. In those cases, we retained the GitHub citation, as no published alternative is currently available. We also clarified these cases explicitly in the response.

- Coherent and incoherent reflection and transmission of multilayer structures: DOI added.
- Thin-Film Optical Filters: DOI added.
- Automatic differentiation in machine learning: a survey: DOI added.
- Multilayer optical calculations: DOI added.
- Transfer matrix method: Python implementation (FilipDominec/transfer_matrix_method): GitHub repository only, no DOI available.
- JAX: composable transformations of Python+NumPy programs: At the time of review, no published paper was available; hence the GitHub repository was cited.
- XLA: Accelerated Linear Algebra Compiler for Machine Learning: GitHub repository only, no DOI available.
- ColorPy – A Python package for handling physical descriptions of color and light spectra: GitHub repository only, no DOI available.
- Optax (DeepMind JAX Ecosystem): We followed the official citation provided in the repository.
- Adam: A Method for Stochastic Optimization: DOI added.
- Autograd: Effortless gradients in NumPy: Workshop paper without DOI; only GitHub repository available.
- General transfer-matrix method for optical multilayer systems with coherent, partially coherent, and incoherent interference: DOI added.
- Optimization of Multilayer Optical Films with a Memetic Algorithm and Mixed Integer Programming: DOI added.
- Refractiveindex.info database of optical constants: DOI added.
- On the limited memory BFGS method for large scale optimization: DOI added.

2. Action taken:
The paper.bib file was updated accordingly to include all available DOIs.
