### 2025-09-30 bahremsd/tmmax#30 @HectorMozo3110
Packaging & Dependencies (JOSS review)
👋@bahremsd

While reviewing the repository, I noticed that it currently includes both a **`pyproject.toml`** and a **`requirements.txt`** file.
Maintaining two separate files for dependencies can cause confusion about which one should be treated as the main source.

Following modern Python packaging practices (PEP 621), it would be cleaner and more transparent to consolidate around `pyproject.toml` as the **authoritative source**, while keeping a `requirements.txt` only when pinned, reproducible environments are strictly needed for CI or deployment.

---

#### Suggested `pyproject.toml` (Flit, PEP 621)

> **Example:** The snippet below shows how dependencies and optional extras could be structured.
> Fields like license, URLs, classifiers, and dependencies should be adjusted to match the real project context.
> If there are no runtime dependencies, the `dependencies` list can remain empty, which is perfectly valid.

```toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "lumache"
authors = [{name = "Bahrem Serhat Danis", email = "bdanis23@ku.edu.tr"}]
dynamic = ["version", "description"]

# Clear metadata improves distribution and transparency
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
keywords = ["example", "template", "packaging"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

# Runtime dependencies
dependencies = [
  "jax>=0.5.0",
  "matplotlib>=3.10.0",
  "numpy>=2.2.2",
  "pandas>=2.2.3",
  "scipy>=1.15.1"
]

[project.optional-dependencies]
# Extras are recommended because they let people install only what they need:
# - tests → run the test suite
# - docs  → build the documentation
# - dev   → contribute with linting, formatting, and type-checking
test = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "coverage[toml]>=7.5",
  "hypothesis>=6.0"
]

docs = [
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
  "mkdocstrings[python]>=0.25",
  "mkdocs-literate-nav>=0.6",
  "mkdocs-section-index>=0.3",
  "mkdocs-gen-files>=0.5",
  "mike>=2.1"
]

dev = [
  "ruff>=0.6",
  "black>=24.0",
  "pre-commit>=3.6",
  "mypy>=1.10",
  "build>=1.2",
  "twine>=5.0"
]

[project.urls]
Homepage = ""
Repository = ""
Issues = ""
Documentation = ""

[tool.coverage.run]
branch = true
source = ["lumache"]

[tool.pytest.ini_options]
addopts = "-q --cov=lumache --cov-report=term-missing"
```

---

#### Quick Start for Users and Collaborators

With this setup, different groups can install exactly what they need:

**Install runtime only:**
```bash
pip install .
```

**Install with tests and run them:**
```bash
pip install .[test]
pytest -q
```

**Install with docs and build the site:**
```bash
pip install .[docs]
mkdocs build
```

**Install everything (for contributors):**
```bash
pip install .[test,docs,dev]
```

---

#### `requirements.txt` for CI/Deployment (Optional)

A `requirements.txt` is not mandatory.
If pinned versions are needed for reproducibility in CI or deployment, it can be generated from the environment:

```bash
pip install .[test,docs,dev]
pip freeze > requirements.txt
```

Otherwise, `pyproject.toml` should be considered the single authoritative source of dependencies.

---

#### Continuous Integration (Optional)

Adding a minimal GitHub Actions workflow would further help by:
- Installing the package (`pip install .[test]`)
- Running the tests (`pytest -q`)
- Optionally building the docs (`pip install .[docs] && mkdocs build`)

----
The current installation approach looks good and works as expected, but it feels a bit limited in scope. Consolidating everything in pyproject.toml with extras for tests, docs, and dev would add more transparency and make the workflow clearer for users and collaborators. This matters because it helps ensure reproducibility and lowers the barrier for people who want to use or contribute to the project. It’s not strictly mandatory, but I believe it would make the setup more complete. ***What do you think about moving in this direction?***

**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/joss-reviews/issues/9088)
