### 2025-09-30 bahremsd/tmmax#31 @HectorMozo3110
Tests & Reproducibility (JOSS Review)
👋 @bahremsd
While checking the repository, I noticed that there is currently a single test file (`tests/test_tmm.py`).
The file only contains very simple arithmetic checks:

```python
def test_addition():
    assert 2 + 2 == 4

def test_subtraction():
    assert 5 - 3 == 2
```

Although this shows that a testing framework is set up, it does not actually validate the functionality of the package itself.
For JOSS, it is important that tests cover the **core features of the software**. According to the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), the submitted software should include tests that can be run automatically and that verify key functionality.

---

#### What could be added
Some examples of useful tests for TMMax include:
- **Import tests**: confirm that the package can be imported without errors.
- **Functionality tests**: check that main functions or classes return the expected outputs for small sample inputs.
- **Integration tests**: validate that a short workflow (e.g., a pipeline or example from the README) runs successfully end-to-end.
- **Error handling tests**: ensure that the package raises clear errors when given invalid inputs.

---

Would you consider expanding the test suite to include these types of checks?
This would make the project stronger, more reproducible, and fully aligned with JOSS review expectations.

**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/joss-reviews/issues/9088)
