### 2025-10-01 bahremsd/tmmax#31 @CrazeXD
@bahremsd @HectorMozo3110 I concur with this feedback - I would also recommend providing your benchmarking tests/simulation code against other softwares that also perform TMM in the repository as part of the testing suite. e.g. the code for figures 2 and 3 as you see fit.
