### 2025-10-05 #9088 @bahremsd
Hello @CrazeXD,

We greatly appreciate your positive feedback in the issue https://github.com/bahremsd/tmmax/issues/27#issuecomment-3353363714. Below, we provide detailed responses and the actions taken for each point you raised in the issue:

---

### 1. Comment:

> I apologize for not mentioning this prior, but as per JOSS guidelines, the target audience should be placed into the Statement of Need. However, I do like the way that you have integrated this into the summary, so if you have the space (that I do acknowledge you have very limited of), leaving it as it is and adding in some detail into the Statement of Need could be beneficial. This is your choice entirely, but the target audience does need to be present in the statement of need.

**Our response:**
We believe that integrating the target audience into the summary strengthens the summary itself. Additionally, we added a sentence at the end of the Statement of Need to indicate the target audience who would find TMMax useful. To maintain the 1000-word limit, we shortened some sentences without losing meaning and moved extra details to the documentation.

**Action taken:**

- **Removed from the Statement of Need:**
  > “Additionally, conventional TMM packages that focus on system matrices and offer limited, manually entered material data, TMMax incorporates a curated database of 30 widely used materials, selected from Thin-Film Optical Filters and sourced from the verified refractiveindex.info database.”

- **Added:**
  > “Additionally, TMMax integrates a curated database of 30 extensively used optical materials, sourced from refractiveindex.info, thereby enabling optical engineers and thin-film researchers in optics and photonics to efficiently simulate complex dielectric multilayer structures through a scalable, JAX-accelerated implementation.”

- **Shortened the benchmark sentence:**
  - Original:
    > “We used Python’s timeit module to benchmark each simulation 50 times, averaging results to minimize transient effects and all comparisons run on a single Intel Core i9 core without GPU or multicore use to ensure fairness.”
  - Updated:
    > “We used Python’s timeit module to benchmark each simulation 50 times, with all comparisons run on a single Intel Core i9 core without GPU or multicore use for fairness.”

---

### 2. Comment:

> A small typo I notice: In line 31, you have a period before the citation "(Byrnes 2020)", which looks to be out of place.

**Our response:**
Thank you for noticing the typo, we have corrected it.

**Action taken:**

- Original:
  > “In traditional TMM implementations, the stack of layers in Figure 1a is simulated using a single wavelength and angle of incidence, as shown in Figure 1b, and nested loops over wavelengths and angles lead to redundant calculations. (Byrnes 2020).”

- Updated:
  > “In traditional TMM implementations, the stack of layers in Figure 1a is simulated using a single wavelength and angle of incidence, as shown in Figure 1b, and nested loops over wavelengths and angles lead to redundant calculations (Byrnes 2020).”

---

### 3. Comment:

> Additionally, do you have a citation for the sentence beginning on line 52 of the paper, or is this something you saw to be true experimentally?

**Our response:**
This is an intrinsic property of TMM. As the number of layers increases, the number of 2x2 matrix multiplications naturally increases, which increases runtime. Similarly, longer wavelength or angle-of-incidence arrays increase the number of transfer matrix multiplications. In conventional TMM, this corresponds to nested loops executing more iterations, whereas in vectorized TMM, the size of the vectorization axis increases. We modified the sentence to reflect that runtime scales with the number of layers, wavelength array length, and angle-of-incidence array length due to the number of matrix multiplications.

**Action taken:**

- Original:
  > “In multilayer thin-film simulations using TMM, runtime critically depends on the number of layers, wavelength array length, and angle of incidence array length, each substantially affecting computational load. To benchmark TMMax, we used Steven Byrnes’ Python tmm library as a baseline.”

- Updated:
  > “Runtime in TMM scales naturally with the number of layers, as well as the lengths of the wavelength and incidence-angle arrays, due to the increased number of transfer matrix multiplications. To benchmark TMMax, we used tmm library as a reference.”
