### 2025-10-05 #9088 @bahremsd
Hello @HectorMozo3110 ,

First, we sincerely thank you for your positive feedback in the issue https://github.com/bahremsd/tmmax/issues/28#issue-3469645763 . Below, we provide structured responses and actions taken for each point you mentioned:

## 1. Installation Section

**Comment:**
Installation section: Since the paper already includes a brief installation instruction, it would be good to also add a direct reference to the official documentation (README or docs). That way, users can go from the paper directly to the full installation guide with dependencies and platform notes.

**Our Answer:**
We added an additional sentence to the Installation section that directly directs readers to the [TMMax Documentation](https://tmmax.readthedocs.io/en/latest/index.html).

**Action Taken:**
We inserted the following sentence in the Installation section:
> "For detailed installation instructions and platform compatibility, please refer to the [TMMax Documentation](https://tmmax.readthedocs.io/en/latest/index.html)."

Additionally, to keep the word count below 1000, we shortened the method description at the end of the benchmark section. The original sentence:
> "We used Python’s timeit module to benchmark each simulation 50 times, averaging results to minimize transient effects and all comparisons run on a single Intel Core i9 core without GPU or multicore use to ensure fairness."

was revised to:
> "We used Python’s timeit module to benchmark each simulation 50 times, with all comparisons run on a single Intel Core i9 core without GPU or multicore use for fairness."

## 2. Demonstrated Use

**Comment:**
Demonstrated use (if applicable): Just as a clarification question: has TMMax already been used in any real research project, lab setting, or conference work? If so, even a short mention in the paper or repo would help show its applicability.

**Our Answer:**
TMMax has been used in a research project; however, the paper of the project has not yet been published. Therefore, we cannot reference it at this time. Once this paper is published, or if other published works using TMMax become available, we will add a section titled "Projects Using TMMax" in the documentation to highlight these applications.

**Action Taken:**
No immediate update was made, as no published projects are currently available to reference. This section will be updated in future documentation releases.

## 3. Equal-Contrib Field

**Reviewer Comment:**
I was also going to open a small PR regarding the equal-contrib field in the author metadata, but I noticed that this option appears to be deactivated. Do you prefer to keep it as it is, or would you prefer that I submit the PR so it can be marked accordingly?

**Our Answer:**
We realized that the "equal-contrib" field was unnecessary. We have removed it from paper.md, as there is no equal contribution in our manuscript.

**Action Taken:**
The "equal-contrib" entries were deleted from paper.md. The updated version now accurately reflects the authorship information without this field.
