### 2025-10-07 bahremsd/tmmax#31 @bahremsd
Dear Reviewers @HectorMozo3110 @CrazeXD ,

We sincerely thank you for your feedback and for highlighting the importance of test coverage. We have significantly expanded and refined the testing suite to ensure full alignment with the JOSS review criteria. Below, we provide a detailed overview of the newly added test files, their purposes:

---

## 1. `test_imports.py`

This test file verifies that all the essential libraries and modules required by **TMMax** can be imported correctly.
It ensures that users can access all top-level functions and that the dependency management defined in `pyproject.toml` is fully functional across different environments.

---

##  2. `test_functions.py`

This file includes functionality tests for two key utility functions:

- `visualize_material_properties()`
- `add_material_to_nk_database()`

---

##  3. `test_integration.py`

This file performs integration tests focusing on the **core function `tmm`**.
Specifically, it evaluates:

- Coherent and incoherent configurations
- Different polarization modes (s- and p-polarization)

---

## 4. `test_errors.py`

This file validates TMMax’s robustness in handling erroneous or unexpected input types.
We tested the core function `tmm` with:

- Arrays of incompatible shapes and lengths
- Incorrect input types
- Invalid material names or polarization parameters

---

##  5. Continuous Integration and Automated Testing

All of the above test files are automatically executed through the `release-package.yaml` workflow in **GitHub Actions**.
Whenever the repository is updated, the tests are triggered automatically to ensure stability and reproducibility.

To replicate the same testing environment locally:

```bash
git clone https://github.com/bahremsd/tmmax.git
pip install .[test]
pytest --ignore=tests/test_benchmark.py --cov=tmmax --cov-report=term-missing
```

---

##  6. Benchmark and Validation Tests — `test_benchmark.py`

In addition to the functional tests, we have also included benchmarking and validation code corresponding to **Figure 2** and **Figure 3** in the manuscript.
These benchmark scripts provide quantitative comparisons of computational performance and accuracy across varying layer counts and array lengths.

To reproduce the benchmark figures:

```bash
git clone https://github.com/bahremsd/tmmax.git
pip install .[test]
pytest tests/test_benchmark.py
```

> **Note:** Benchmark runs are computationally intensive.
> - For *Figure 2 (Layer Count vs. Time)* → up to **50 layers**, repeated **10 times**
> - For *Figure 3 (Array Length Benchmark)* → array lengths up to **22**, repeated **3 times**
> Total runtime: ~30 minutes.

The figures below were generated from our internal benchmark run using these configurations and confirm the reproducibility of the published results.

---

8 layer :

80 layer:
