### 2025-10-07 bahremsd/tmmax#30 @bahremsd
Dear reviewer @HectorMozo3110 ,

We sincerely thank you for their insightful and constructive feedback regarding the repository’s **dependency management structure**.   We fully agree that consolidating around a single authoritative configuration file is a best practice aligned with modern Python packaging standards, particularly those described in **PEP 621** and the **Flit build system**.

---

## Repository Configuration Update

Following this valuable suggestion, we have carefully revised our packaging configuration to establish `pyproject.toml` as the **canonical and comprehensive source** for build, dependency, and metadata information.
The previously included `requirements.txt` file has been **deprecated** in favor of this unified configuration, ensuring greater **transparency**, **maintainability**, and **reproducibility** of the environment.

---

## 📦 Updated `pyproject.toml`

```toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "tmmax"
authors = [{name = "Bahrem Serhat Danis", email = "bdanis23@ku.edu.tr"}]
dynamic = ["version", "description"]

readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["tmm", "optics", "photonics", "jax", "thin", "film", "dielectric", "vectorization", "jit", "coating"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "jax>=0.5.0",
  "matplotlib>=3.10.0",
  "numpy>=2.2.2",
  "pandas>=2.2.3",
  "scipy>=1.15.1"
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "coverage[toml]>=7.5",
  "hypothesis>=6.0",
  "tmm==0.2.0",
  "matplotlib",
]

docs = [
  "sphinx==7.1.2",
  "sphinx-rtd-theme==1.3.0rc1",
  "sphinx-favicon==1.0.1",
  "furo==2025.7.19",
  "sphinxcontrib-bibtex",
  "nbsphinx",
  "nbconvert",
  "ipykernel"
]

dev = [
  "ruff>=0.6",
  "black>=24.0",
  "pre-commit>=3.6",
  "mypy>=1.10",
  "build>=1.2",
  "twine>=5.0"
]

[project.urls]
Homepage = "https://github.com/bahremsd/tmmax"
Repository = "https://github.com/bahremsd/tmmax"
Issues = "https://github.com/bahremsd/tmmax/issues"
Documentation = "https://tmmax.readthedocs.io/en/latest/"

[tool.coverage.run]
branch = true
source = ["tmmax"]

[tool.pytest.ini_options]
addopts = "-q --cov=tmmax --cov-report=term-missing"
```

---
