### 2025-10-08 bahremsd/tmmax#31 @HectorMozo3110
@bahremsd  Thank you for the detailed update. The new tests and benchmarks look good to me. I’ll run a few additional tests over the next few days, and with that, I’ll finalize my review. For now, everything looks great. Thanks.

@CrazeXD what do you think about this new update? Do you think we can consider this point closed, or would you suggest any further improvements?
