### 2025-10-11 b2b0ac1
Update paper.md
