### 2025-10-22 #9088 @lockwo
Can you update the title/author list in the metadata of the archived release to match the paper @bahremsd ? The title should exactly match the paper title (and the author list of the paper should be the author list in the metadata). I don't believe this requires a new zenodo release, so I'll go ahead and tag the versions for JOSS.
