### 2025-10-22 #9088 @bahremsd
Sure! @lockwo We have updated the title/author list in the metadata of the archived release:  [https://zenodo.org/records/17361547](https://zenodo.org/records/17361547)
