### 2025-10-23 #9088 @lockwo
Just a few minor grammar things in the paper I'm noticing on final check, then I should be good to recommend accept.

- "TMMax supports CPU, GPU, and TPU hardware, includes" -> TMMax supports CPU, GPU, and TPU hardware, and includes
- "subsequently apply the JAX's" -> subsequently apply JAX's
