### 2025-10-24 #9088 @kyleniemeyer
@lockwo paper-only updates do not require updates to the archive, only changes to the software itself (the paper is not part of the software / version, and we archive the paper)
